FLEXY-1681-i bug: importing invokes command
