+++ packages/flex-plugin-scripts/src/scripts/info.ts
@@ -2,26 +2,7 @@ import { logger } from 'flex-dev-utils';
 import { packagesVersions } from '../prints';
 
 import run from '../utils/run';
-import { getPackageDetails } from '../utils/package';
-
-export const FLEX_PACKAGES: string[] = [
-  '@twilio/flex-ui',
-  'flex-plugin-scripts',
-  'flex-plugin',
-  'flex-dev-utils',
-  'craco-config-flex-plugin',
-];
-
-/* istanbul ignore next */
-const LIST_OF_PACKAGES: string[] = [
-  ...FLEX_PACKAGES,
-  '@craco/craco',
-  'react-scripts',
-  'react',
-  'react-dom',
-  'redux',
-  'react-redux',
-];
+import { getPackageDetails, LIST_OF_PACKAGES } from '../utils/package';
 
 /**
  * Prints the list of versions of important packages
